+++ src/App.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { RouterView } from 'vue-router'
-import HeaderComponent from '@/components/header/Header-component.vue'
+import HeaderComponent from '@/components/header/HeaderComponent.vue'
 import PromoBanner from './components/PromoBanner/PromoBanner.vue'
 import FooterComponent from './components/FooterComponent/FooterComponent.vue'
 import Toast from 'primevue/toast'
